move plugin into folder
