+++ lib/WebpackOptionsApply.js
@@ -21,7 +21,7 @@ const SourceMapDevToolPlugin = require("./SourceMapDevToolPlugin");
 const EntryOptionPlugin = require("./EntryOptionPlugin");
 const RecordIdsPlugin = require("./RecordIdsPlugin");
 
-const InferAsyncModulesPlugin = require("./InferAsyncModulesPlugin");
+const InferAsyncModulesPlugin = require("./async-modules/InferAsyncModulesPlugin");
 const RuntimePlugin = require("./RuntimePlugin");
 
 const APIPlugin = require("./APIPlugin");
