+++ lib/async-modules/InferAsyncModulesPlugin.js
@@ -5,12 +5,12 @@
 
 "use strict";
 
-const HarmonyImportDependency = require("./dependencies/HarmonyImportDependency");
-const HarmonyImportSideEffectDependency = require("./dependencies/HarmonyImportSideEffectDependency");
-const WebpackError = require("./WebpackError");
+const HarmonyImportDependency = require("../dependencies/HarmonyImportDependency");
+const HarmonyImportSideEffectDependency = require("../dependencies/HarmonyImportSideEffectDependency");
+const WebpackError = require("../WebpackError");
 
-/** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./Module")} Module */
+/** @typedef {import("../Compiler")} Compiler */
+/** @typedef {import("../Module")} Module */
 
 class InferAsyncModulesPlugin {
 	constructor(options) {
